pyproject.toml
[project]
name = "site-log"
version = "0.1.0"
description = "Geofenced time-on-site logger."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "django-leaflet>=0.32.0",
    "djangorestframework>=3.16.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "basedpyright>=1.31.4",
    "django-stubs>=5.2.2",
    "ruff>=0.12.12",
]

[tool.ruff]
respect-gitignore = true
exclude = [
  "__pycache__",
  ".ruff_cache",
  "./geologger/migrations",
]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "UP",
  "B",
  "SIM",
  "I",
]
ignore = [
  "SIM108",  # Don't enforce ternary operator
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.basedpyright]
include = ["geologger", "site_log"]
exclude = [
  ".venv",
  "**/__pycache__",
  "**/.ruff_cache",
  "**/tests",
]
pythonVersion = "3.13"
reportMissingImports = "error"
